Separate containers from components, refactor App component
